+++ CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Security
 
+## [7.1.2](https://github.com/ruflin/Elastica/compare/7.1.1...7.1.2)
+### Backward Compatibility Breaks
+* Changed `\Elastica\Suggest\CandidateGenerator\DirectGenerator::toArray` to return a list of params [#1964](https://github.com/ruflin/Elastica/pull/1964)
+### Added
+* Added `\Elastica\Suggest\Phrase::addDirectGenerator` to align with ES specification [#1964](https://github.com/ruflin/Elastica/pull/1964)
+
 ## [7.1.1](https://github.com/ruflin/Elastica/compare/7.1.0...7.1.1)
 ### Backward Compatibility Breaks
 * Changed `Elastica\Query\MatchQuery::setFieldParam()` signature to allow passing bool, float or int [#1941](https://github.com/ruflin/Elastica/pull/1941)